Remove unnecessary context for channel because we assume non-blocking startTask
Reorganize some files
